models/ai-agent/requirements.txt
accelerate==1.1.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.4
Deprecated==1.2.14
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
fastapi==0.115.5
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
guidance==0.1.13
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.6
interegular==0.3.3
Jinja2==3.1.3
joblib==1.3.2
llama-hub==0.0.79.post1
llama-index==0.11.23
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.23.post2
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.2
llama-index-legacy==0.9.48
llama-index-llms-huggingface==0.1.4
llama-index-llms-llama-cpp==0.2.3
llama-index-llms-ollama==0.3.6
llama-index-llms-openai==0.2.16
llama-index-multi-modal-llms-openai==0.2.3
llama-index-program-guidance==0.2.1
llama-index-program-lmformatenforcer==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.3.0
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.14
llama_cpp_python==0.3.1
llamaindex-py-client==0.1.15
lm-format-enforcer==0.10.9
MarkupSafe==3.0.2
marshmallow==3.21.1
minijinja==2.5.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==2.1.3
openai==1.14.3
ordered-set==4.1.0
packaging==24.0
pandas==2.2.1
pillow==11.0.0
platformdirs==4.2.0
protobuf==5.26.1
psutil==6.1.0
pyaml==24.9.0
pydantic==2.6.4
pydantic_core==2.16.3
pydot==3.0.2
pyformlang==1.0.9
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
retrying==1.3.4
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==3.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.41.2
striprtf==0.0.27
sympy==1.12
tenacity==9.0.0
threadpoolctl==3.4.0
tiktoken==0.8.0
tokenizers==0.20,3
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.32.0
wrapt==1.16.0
yarl==1.9.4